Добавляет маршрутизацию.
