+++ project/src/components/app/app.tsx
@@ -1,4 +1,12 @@
-import PageMain from '../page-main/page-main';
+import {Route, BrowserRouter, Routes} from 'react-router-dom';
+import {AppRoute, AuthorizationStatus} from '../../const';
+import PrivateRoute from '../private-route/private-route';
+import Layout from '../layout/layout';
+import Main from '../../pages/main/main';
+import NotFound from '../../pages/not-found/not-found';
+import Login from '../../pages/login/login';
+import Favorites from '../../pages/favorites/favorites';
+import Room from '../../pages/room/room';
 
 type AppProps = {
   rentOffersCount: number;
@@ -6,7 +14,24 @@ type AppProps = {
 
 function App({rentOffersCount}: AppProps): JSX.Element {
   return (
-    <PageMain rentOffersCount={rentOffersCount} />
+    <BrowserRouter>
+      <Routes>
+        <Route path='/' element={<Layout/>}>
+          <Route index element={<Main rentOffersCount={rentOffersCount}/>}/>
+          <Route path={AppRoute.Login} element={<Login/>}/>
+          <Route path={AppRoute.Favorites}
+            element={
+              <PrivateRoute authorizationStatus={AuthorizationStatus.NoAuth}>
+                <Favorites />
+              </PrivateRoute>
+            }
+          />
+          <Route path={AppRoute.Room} element={<Room/>}/>
+          <Route path="*" element={<NotFound/>}/>
+        </Route>
+      </Routes>
+    </BrowserRouter>
+
   );
 }
