+++ project/src/components/layout/layout.tsx
@@ -1,0 +1,35 @@
+import {Fragment} from 'react';
+import {Link, Outlet} from 'react-router-dom';
+import Icons from '../icons/icons';
+
+function Layout(): JSX.Element {
+  return (
+    <Fragment>
+      <Icons/>
+      <div className="page page--gray page--main">
+        <header className="header">
+          <div className="container">
+            <div className="header__wrapper">
+              <div className="header__left">
+                <Link className="header__logo-link" to="/">
+                  <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41"/>
+                </Link>
+              </div>
+            </div>
+          </div>
+        </header>
+
+        <main className="page__main page__main--index">
+          <Outlet/>
+        </main>
+
+        <footer className="footer container">
+          <Link className="footer__logo-link" to="/">
+            <img className="footer__logo" src="img/logo.svg" alt="6 cities logo" width="64" height="33"/>
+          </Link>
+        </footer>
+      </div>
+    </Fragment>);
+}
+
+export default Layout;
