+++ project/src/const.ts
@@ -1,0 +1,11 @@
+export enum AppRoute {
+  Login = '/login',
+  Favorites = '/favorites',
+  Room = '/offer/:id'
+}
+
+export enum AuthorizationStatus {
+  Auth = 'AUTH',
+  NoAuth = 'NO_AUTH',
+  Unknown = 'UNKNOWN',
+}
