+++ project/src/pages/login/login.tsx
@@ -1,0 +1,29 @@
+function Login(): JSX.Element {
+  return (
+    <div className="page__login-container container">
+      <section className="login">
+        <h1 className="login__title">Sign in</h1>
+        <form className="login__form form" action="#" method="post">
+          <div className="login__input-wrapper form__input-wrapper">
+            <label className="visually-hidden">E-mail</label>
+            <input className="login__input form__input" type="email" name="email" placeholder="Email" required/>
+          </div>
+          <div className="login__input-wrapper form__input-wrapper">
+            <label className="visually-hidden">Password</label>
+            <input className="login__input form__input" type="password" name="password" placeholder="Password" required/>
+          </div>
+          <button className="login__submit form__submit button" type="submit">Sign in</button>
+        </form>
+      </section>
+      <section className="locations locations--login locations--current">
+        <div className="locations__item">
+          <a className="locations__item-link" href="#">
+            <span>Amsterdam</span>
+          </a>
+        </div>
+      </section>
+    </div>
+  );
+}
+
+export default Login;
